api/requirements.txt
Flask==3.1.1
cohere==5.15.0
flask-cors==6.0.1
requests==2.32.4
python-dotenv==1.1.0